Add initial Morello machines and layer config
The machines are split between bsp/linux musl-only/linux with glibc.
This way the users will be able to build a purecap-musl-only environment in the
future or a glibc environment with purecap being statically linked. The bsp is its
own separate machine to deal with the fact that LLVM Morello is not a fully
implemented toolchain (we might get build errors relating to gcc lib which
needs libc, but libc is purecap only).

Signed-off-by: Pawel Zalewski <pzalewski@thegoodpenguin.co.uk>
Reviewed-by: Andrew Murray <amurray@thegoodpenguin.co.uk>